polish: fix stale comments, drop dead env assignments, dedupe test double
