+++ src/node/services/hooks.test.ts
@@ -12,31 +12,7 @@ import {
   runPostHook,
 } from "./hooks";
 import { LocalRuntime } from "@/node/runtime/LocalRuntime";
-import type { ExecOptions, ExecStream } from "@/node/runtime/Runtime";
-
-class ExecPathMappingRuntime extends LocalRuntime {
-  constructor(
-    projectPath: string,
-    private readonly hostPrefix: string,
-    private readonly execPrefix: string
-  ) {
-    super(projectPath);
-  }
-
-  override exec(command: string, options: ExecOptions): Promise<ExecStream> {
-    const mapPath = (filePath: string) =>
-      filePath.startsWith(this.hostPrefix)
-        ? this.execPrefix + filePath.slice(this.hostPrefix.length)
-        : filePath;
-    return super.exec(command, {
-      ...options,
-      cwd: mapPath(options.cwd),
-      pathEnv: Object.fromEntries(
-        Object.entries(options.pathEnv ?? {}).map(([key, value]) => [key, mapPath(value)])
-      ),
-    });
-  }
-}
+import { ExecPathMappingRuntime } from "./testExecPathMappingRuntime";
 
 describe("hooks", () => {
   let tempDir: string;
